requirements.txt
# Data ingestion
requests>=2.31.0
pyarrow>=14.0.0
pandas>=2.1.0

# Database
psycopg2-binary>=2.9.9

# dbt
dbt-core>=1.7.0
dbt-postgres>=1.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# Streaming (producer only — PySpark consumer uses .streaming-venv with Python 3.13)
confluent-kafka>=2.3.0

# Data quality
# GE requires Python <=3.12 (Pydantic V1 incompatible with 3.13+).
# In this project GE runs inside the Airflow container (Python 3.11).
# To run scripts/run_ge.py locally, use a Python 3.11 or 3.12 venv.
great-expectations==0.18.22
sqlalchemy>=2.0,<3.0
jinja2>=3.1.0          # also pulled by dbt; pin explicitly for run_ge.py
